+++ spec/lib/sidekiq/throttled/basic_fetch_spec.rb
@@ -1,0 +1,84 @@
+# frozen_string_literal: true
+
+require "sidekiq/throttled/basic_fetch"
+
+class ThrottledTestJob
+  include Sidekiq::Job
+  include Sidekiq::Throttled::Job
+
+  def perform(*); end
+end
+
+RSpec.describe Sidekiq::Throttled::BasicFetch do
+  subject(:fetch) do
+    if Gem::Version.new(Sidekiq::VERSION) < Gem::Version.new("7.0.0")
+      Sidekiq.instance_variable_set(:@config, Sidekiq::DEFAULTS.dup)
+      Sidekiq.queues = %w[default]
+      described_class.new(Sidekiq)
+    else
+      config = Sidekiq::Config.new
+      config.queues = %w[default]
+      described_class.new(config.default_capsule)
+    end
+  end
+
+  describe "#retrieve_work" do
+    def enqueued_jobs(queue)
+      Sidekiq.redis do |conn|
+        conn.lrange("queue:#{queue}", 0, -1).map do |job|
+          JSON.parse(job).then do |payload|
+            [payload["class"], payload["args"]]
+          end
+        end
+      end
+    end
+
+    before do
+      # Sidekiq is FIFO queue, with head on right side of the list,
+      # meaning jobs below will be stored in 3, 2, 1 order.
+      ThrottledTestJob.perform_async(1)
+      ThrottledTestJob.perform_async(2)
+      ThrottledTestJob.perform_async(3)
+    end
+
+    context "when job is not throttled" do
+      it "returns unit of work" do
+        expect(fetch.retrieve_work).to be_an_instance_of(Sidekiq::BasicFetch::UnitOfWork)
+      end
+    end
+
+    context "when job was throttled due to concurrency" do
+      before do
+        ThrottledTestJob.sidekiq_throttle(concurrency: { limit: 1 })
+        fetch.retrieve_work
+      end
+
+      it "returns nothing" do
+        expect(fetch.retrieve_work).to be_nil
+      end
+
+      it "pushes job back to the end queue" do
+        expect { fetch.retrieve_work }
+          .to change { enqueued_jobs("default") }
+          .to contain_exactly(["ThrottledTestJob", [2]], ["ThrottledTestJob", [3]])
+      end
+    end
+
+    context "when job was throttled due to threshold" do
+      before do
+        ThrottledTestJob.sidekiq_throttle(threshold: { limit: 1, period: 60 })
+        fetch.retrieve_work
+      end
+
+      it "returns nothing" do
+        expect(fetch.retrieve_work).to be_nil
+      end
+
+      it "pushes job back to the end queue" do
+        expect { fetch.retrieve_work }
+          .to change { enqueued_jobs("default") }
+          .to contain_exactly(["ThrottledTestJob", [2]], ["ThrottledTestJob", [3]])
+      end
+    end
+  end
+end