+++ spec/support/capybara.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
 require "capybara/rspec"
-require "capybara/apparition"
-
-Capybara.javascript_driver = :apparition
 
 RSpec.configure do |config|
   config.before(type: :feature) { Capybara.app = Sidekiq::Web }